requirements.txt
Flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.0
reportlab==4.0.4